// src/app/shopping-list/model/items.ts
export interface Items {
    active: Item;
    list: Item[];
  }
  
  export interface Item {
    id?: string;
    item?: string;
    amount?: string;
  }
  
  // export interface Users {
  //   active: User;
  //   list: User[];
  // }
  
  // export interface User {
  //   id?: number;
  //   name?: string;
  // }
  